+++ gmate/__init__.py
@@ -1,0 +1,5 @@
+from flask import Flask 
+
+app = Flask(__name__)
+
+from gmate import routes
