+++ gmate/templates/connectionError.html
