+++ html_attribution.md
@@ -1,0 +1,6 @@
+# Login
+ _A Pen created at CodePen.io. Original URL: [https://codepen.io/Ri2parna/pen/abbzWaQ](https://codepen.io/Ri2parna/pen/abbzWaQ).
+
+ Inspired by <a href="http://dribbble.com/shots/975425-Flat-UI-login">Virgil Pana</a>
+
+Source Code on GitHub: https://github.com/marcobiedermann/playground/tree/master/ui/login/login
