--- requirements.txt
@@ -1,3 +1,0 @@
-Flask==1.1.1
-Jinja2==2.10.3
-PyGithub==1.43.8
