--- templates/index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-	<title>index</title>
-	<link rel="stylesheet" type="text/css" href="/static/style.css">
-</head>
-<body>
-	<div>
-	<h1>GitHub Mate</h1>
-	<img src="/static/assets/github.png">
-	<form method="post" action="/profile">
-		<input type="text" name="username" placeholder="Username">
-		<br> <br>
-		<button type="submit">Submit</button>
-	</form>
-</div>
-</body>
-</html>
