--- templates/profile.html
@@ -1,58 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-	<title>profile</title>
-	<link rel="stylesheet" type="text/css" href="/static/style.css">
-</head>
-<body class="profile">
-
-	{% if error %}
-	<p class="alert">User doesn't exist!</p>
-
-	{% else %}
-	<img class="img" src="{{ user.avatar_url }}">
-	<h1 class="name">{{ name }}</h1>
-	<ul class="user-details">
- master
-		<li><a href="https://github.com/{{username}}/repositories" target="_blank">Repositories: <span>{{ repos }}</span></a></li>
-		<li><a href="https://github.com/{{username}}/followers" target="_blank">Followers: <span>{{ followers }}</span></a></li>
-		<li><a href="https://github.com/{{username}}/following" target="_blank">Following: <span>{{ following }}</span></a></li>
-		<li><a href="https://gist.github.com/{{username}}" target="_blank">Public Gists: <span>{{ gists }}</span></a></li>
-		<!--<li><a>Repositories with Open issues</a><li>-->
-		<!--{% if repos_with_issues %}-->
-		<!--{% for repo_link in repos_with_issues %}-->
-		<!--<li><a href="{{ repo_link }}" target="_blank">No of open issues<span>{{ repos_with_issues.get(repo_link) }}</span></a></li>-->
-		<!--{% endfor %}-->
-		<!--{% else %}-->
-		<!--<li><a href="https://github.com/{{username}}/repositories">No repository with open issues</a></li>-->
-		<!--{% endif %}-->
-
-		<li><a href="https://github.com/{{ user.login }}/repositories" target="_blank">Repositories: <span>{{ user.public_repos }}</span></a></li>
-		<li><a href="https://github.com/{{ user.login }}/followers" target="_blank">Followers: <span>{{ user.followers }}</span></a></li>
-		<li><a href="https://github.com/{{ user.login }}/following" target="_blank">Following: <span>{{ user.following }}</span></a></li>
-		<li><a href="https://gist.github.com/{{ user.login }}" target="_blank">Public Gists: <span>{{ user.public_gists }}</span></a></li>
- master
-	</ul>
-	<br>
-	<br>
-	<ul class="user-details" style="width: 90%">
-	<li>
-		<a style="color:#161c26;">----------------------------Repositories with Open issues----------------------------</a></li>
-		{% if repos_with_issues %}
-		{% for repo_link in repos_with_issues %}
-		<li><a href="{{ repo_link }}" target="_blank">{{ repos_with_issues_name.get(repo_link)}}<span>{{ repos_with_issues.get(repo_link) }}</span></a></li>
-		{% endfor %}
-		{% else %}
-		<li><a href="https://github.com/{{username}}/repositories">No repository with open issues</a></li>
-		{% endif %}
-		</ul>
-	{% endif %}
-	{% if connectionError %}
-	<p>Check your internet connection!</p>
-	{% endif %}
-
-</body>
-</html>
